getter/requirements.txt
aniso8601==3.0.2
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
Flask-JWT-Extended==3.13.0
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
uWSGI==2.0.17.1
Werkzeug==0.14.1
